Xml Subs

+++ dide/util/xml.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 from dide.models import Permanent, PaymentReport, RankCode
+from dide.models import Employee, NonPermanent
 from django.db import connection, transaction
 from lxml import etree
 from time import time
@@ -9,8 +10,10 @@ import datetime
 def rmv_nsp(node):  # function to remove the namespace from node
     return node.tag.rsplit('}', 1)[-1]
 
+
 def read(file):
-        payment_category_fields = ['type', 'startDate', 'endDate', 'month', 'year']
+        payment_category_fields = ['type', 'startDate', 'endDate',
+                                   'month', 'year']
 
         def query_value(field, value):
             if field in ['startDate', 'endDate']:
@@ -37,47 +40,75 @@ def read(file):
             year = el[0].get('year').rsplit('+', 1)[0]
             el = i.xpath('./xs:periodType', namespaces={'xs': ns})
             paytype = el[0].get('value')
-
+            if int(paytype) not in [1, 11, 12, 13]:
+                month = 16
         e = element.xpath('//xs:psp/xs:body/xs:organizations' +
                           '/xs:organization/xs:employees/xs:employee',
                           namespaces={'xs': ns})
 
-        reports = PaymentReport.objects.filter(pay_type=paytype,
-                                               type=month, year=year).count()
+        if int(paytype) in [1, 11]:
+            reports = PaymentReport.objects.filter(pay_type=paytype,
+                                                   type=month,
+                                                   year=year).count()
+            print 'Salary: Found %s records from previous XML file.' % reports
 
-        print 'Found %s records from a previous XML file.' % reports
-
-        if reports > 0:
-            pr = PaymentReport.objects.filter(pay_type=paytype,
-                                              type_id=month, year=year)
-            pr.delete()
-            reports = ''
+            if reports > 0:
+                pr = PaymentReport.objects.filter(pay_type=paytype,
+                                                  type_id=month, year=year)
+                pr.delete()
+                reports = ''
+        else:
+            print 'Importing new XML file or XML file other than salary.'
 
         cursor = connection.cursor()
+        objects = Employee.objects.all()
+        e_dic = {o.vat_number: o for o in objects}
         objects = Permanent.objects.all()
+        p_dic = {o.registration_number: o for o in objects}
+        objects = NonPermanent.objects.all()
+        s_dic = {o.parent_id: o for o in objects}
         ranks = RankCode.objects.all()
-        dic = {o.registration_number: o for o in objects}
-        rankdic = {o.id: o.rank for o in ranks}
+        rankdic = {o.id for o in ranks}
         for i in e:
+            employeeID = 0
+            initrank = 0
             iban = ''
             netAmount1 = ''
             netAmount2 = ''
             rank = 0
             cntr1 += 1
             el = i.xpath('./xs:identification/xs:amm', namespaces={'xs': ns})
-            payemp = dic.get(el[0].text)
-            if payemp:
+            payemp = p_dic.get(el[0].text)
+            if not payemp:
+                print "Registration not found. Trying vat number. "
+                el = i.xpath('./xs:identification/xs:tin',
+                             namespaces={'xs': ns})
+                payemp = e_dic.get(el[0].text)
+                if payemp:
+                    employeeID = payemp.id
+            else:
+                employeeID = payemp.parent.id
+                initrank = payemp.rank_id()
+                print initrank
+            if employeeID != 0:
                 cntr2 += 1
                 el = i.xpath('./xs:identification/xs:bankAccount',
                              namespaces={'xs': ns})
                 iban = el[0].get('iban')
                 el = i.xpath('./xs:identification/xs:scale/xs:rank',
-                             namespaces={'xs': ns})
+                                namespaces={'xs': ns})
                 if el:
-                    rank = 'NULL' if el[0].text not in rankdic else el[0].text
-                if not rank:
-                    rank = payemp.rank_id() or 'NULL'
-
+                    if el[0].text not in rankdic:
+                        rank = 'NULL'
+                    else:
+                        rank = el[0].text
+                    if rank == 'NULL':
+                        if initrank != 0 and initrank is not None:
+                            rank = initrank
+                        else:
+                            rank = 'NULL'
+                print rank
+                import pdb;pdb.set_trace()
                 el = i.xpath('./xs:payment/xs:netAmount1',
                              namespaces={'xs': ns})
                 netAmount1 = el[0].get('value')
@@ -87,7 +118,7 @@ def read(file):
                 sql += "insert into dide_paymentreport (id, employee_id, "
                 sql += "type_id, year, pay_type, rank_id, iban, "
                 sql += "net_amount1, net_amount2) values (NULL, "
-                sql += "%s, %s, %s, %s, %s, " % (payemp.parent.id,
+                sql += "%s, %s, %s, %s, %s, " % (employeeID,
                                                  month,
                                                  year,
                                                  paytype,
@@ -138,7 +169,8 @@ def read(file):
                             cursor.execute(s_s)
                         except:
                             print s_s
-                            
+                            raise
+
                 sql = ''
             else:
                 print el[0].text + " not found in database."
